ExprDiff

Description

This module uses the Expr datatype, and UniversalNum class. It contains a type class and instances for differentiable expressions, and can conduct evaluation, simplicication, and partial differentiation to the differentiable expressions.

The instances contain Double, Float, Int, Integer 4 basic numerical types.

Class DiffExpr: Differentiable Expression

class Eq a => DiffExpr a where

This class has methods over the Expr datatype that assist with construction and evaluation of differentiable expressions

Minimal complete definition

eval, simp, partDiff

Methods

eval :: Map String a -> Expr a -> a

takes a dictionary of variable identifiers and values, and uses it to compute the Expr fully

simp :: Map String a -> Expr a -> Expr a

takes a possible incomplete dictionary and uses it to reduce Expr as much as possible

partDiff :: String -> Expr a -> Expr a

given a var identifier, differentiate IN TERMS of that identifier

simplify :: Map String a -> Expr a -> Expr a

simplify the expression recursively untill it can't be reduced anymore

>>> simplify (Map.fromList [("x",2)]) $ (val 2 !+ var "x") !+ (val 4 !+ var "y")
(val 8) + (var "y")

(!+) :: Expr a -> Expr a -> Expr a

Binary Addition & simplification

(!*) :: Expr a -> Expr a -> Expr a

Binary Miltiplication & simplification

(!^) :: Expr a -> a -> Expr a

Exponentiation & simplification

loga :: a -> Expr a -> Expr a

Logarithm of arbitory base & simplification

sine :: Expr a -> Expr a

Unary Sin op & simplification

cosi :: Expr a -> Expr a

Unary Cos & simplification

ln :: Expr a -> Expr a

Unary Ln & simplification

val :: a -> Expr a

Wrapper for simple values

var :: String -> Expr a

String identifier for variables