ExprParser

Description

This module uses the Expr datatype. The parsers in this module take a string of format and parse it into an expression of corresponding type (i.e Int, Float, Double...)

Documentation

parserExprD

Arguments

:: String
takes a string

-> Expr Double
parses it into a Expr of Double

parserExprF

Arguments

:: String
takes a string

-> Expr Float
parses it into a Expr of Float

parserExprI

Arguments

:: String
takes a string

-> Expr Int
parses it into a Expr of Int

Note: As longs as meet dot, the parsing stops instead of returning an error, so please choose correct parserExpr*

parserExprIg

Arguments

:: String
takes a string

-> Expr Integer
parses it into a Expr of Integer

Note: As longs as meet dot, the parse stops instead of return an eroor, so please choose correct parserExpr*

parserVectorExprD

Arguments

:: String
takes a string

-> Expr [Double]
parses it into a Expr of [Double]

parserVectorExprF

Arguments

:: String
takes a string

-> Expr [Float]
parses it into a Expr of [Float]

parserVectorExprI

Arguments

:: String
takes a string

-> Expr [Int]
parses it into a Expr of [Int]

parserVectorExprIg

Arguments

:: String
takes a string

-> Expr [Integer]
parses it into a Expr of [Integer]