ExprTest

Description

This module uses the Expr datatype.

It contains sample expressions and several properties for testing functions defined in DiffExpr and VectorSpace datatype.

To use QuickCheck, you could simply just type `quickCheck property_fun_name` in the editor, e.g.

>>> quickCheck testD1
+++ OK, passed 100 tests.

Test Properties & Cases for Differentiable Expr

1. Test for Expr Int

sampleExprI1 :: Expr Int

a sample expression over Int

x + y

sampleExprI2 :: Expr Int

a sample expression over Int

x * y

listToExprI :: [Int] -> Expr Int

generate a list of Int for quickcheck over `Expr Int`

testI1 :: Int -> Bool

Check if the expression satisfies:

a + (-a) = 0

testI2 :: [Int] -> Bool

Check if eval works well on adding multiple constants together

testI3 :: [Int] -> Bool

Check if eval satisfies:

expr * 0 = 0

testI4 :: [Int] -> Bool

Check if simplify could add up all the contants together

testI5 :: [Int] -> Bool

Check if simplify could reduce an expression after timing 0 to 0

testI6 :: Int -> Bool

check the property:

d(x ^ 2 + constant)/dx  = d(x ^ 2)/ dx

testI7 :: Int -> Bool

check the property: partial derivative of a non-existing variable of an expr is 0

d(f(x,y))/ dz = 0

testI8 :: Int -> Int -> Int -> Bool

check the property:

(a ^ b) * (a ^ c) = a ^ (b+c)

2. Test for Expr Integer

sampleExprIg1 :: Expr Integer

a sample expression over Integer

x + y

sampleExprIg2 :: Expr Integer

a sample expression over Integer

x * y

listToExprIg :: [Integer] -> Expr Integer

generate a list of Integer for quickcheck over `Expr Int`

testIg1 :: Integer -> Bool

Check if the expression satisfies:

a + (-a) = 0

testIg2 :: [Integer] -> Bool

Check if eval works well on adding multiple constants together

testIg3 :: [Integer] -> Bool

Check if eval satisfies:

expr * 0 = 0

testIg4 :: [Integer] -> Bool

Check if simplify could add up all the contants together

testIg5 :: [Integer] -> Bool

Check if simplify could reduce an expression after timing 0 to 0

testIg6 :: Integer -> Bool

check the property:

d(x ^ 2 + constant)/dx  = d(x ^ 2)/ dx

testIg7 :: Integer -> Bool

check the property: partial derivitave of a non-existing variable of an expr is 0

d(f(x,y))/ dz = 0

testIg8 :: Integer -> Integer -> Integer -> Bool

check the property:

(a ^ b) * (a ^ c) = a ^ (b+c)

3. Test for Expr Double

sampleExprD1 :: Expr Double

sampleExprD2 :: Expr Double

listToExprD :: [Double] -> Expr Double

generate a list of Double for quickcheck over (Expr Double)

testD1 :: Double -> Bool

Check if the expression satisfies:

a + (-a) = 0

testD2 :: [Double] -> Bool

Check if eval workes well on adding multiple constants together

testD3 :: [Double] -> Bool

Check if eval satisfies:

expr * 0 = 0

testD4 :: [Double] -> Bool

Check if simplify could add up all the contants together

testD5 :: [Double] -> Bool

Check if simplify could reduce an expression after timing 0 to 0

testD6 :: Double -> Bool

check the property:

d(x ^ 2 + constant)/dx  = d(x ^ 2)/ dx

testD7 :: Double -> Bool

check the property: partial derivitave of a non-existing variable of an expr is 0

d(f(x,y))/ dz = 0

testD8 :: Double -> Double -> Double -> Bool

check the property:

(a ^ b) * (a ^ c) = a ^ (b+c)

4. Test for Expr Float

sampleExprF1 :: Expr Float

sampleExprF2 :: Expr Float

listToExprF :: [Float] -> Expr Float

generate a list of Double for quickcheck over (Expr Double)

testF1 :: Float -> Bool

Check if the expression satisfies:

a + (-a) = 0

testF2 :: [Float] -> Bool

Check if evalworkes well on adding multiple constants together

testF3 :: [Float] -> Bool

Check if eval satisfies:

expr * 0 = 0

testF4 :: [Float] -> Bool

Check if simplify could add up all the contants together

testF5 :: [Double] -> Bool

Check if simplify could reduce an expression after timing 0 to 0

testF6 :: Float -> Bool

check the property:

d(x ^ 2 + constant)/dx  = d(x ^ 2)/ dx

testF7 :: Float -> Bool

check the property: partial derivitave of a non-existing variable of an expr is 0

d(f(x,y))/ dz = 0

testF8 :: Float -> Float -> Float -> Bool

check the property:

(a ^ b) * (a ^ c) = a ^ (b+c)

Test Properties & Cases for Vector Expr

1. Test for Expr [Int]

sampleVectorExprI1 :: Expr [Int]

a sample vector expression over [Int]

x + y

sampleVectorExprI2 :: Expr [Int]

a sample vector expression over [Int]

x * y

testVI1 :: Int -> Int -> Int -> Bool

Check if the expression satisfies:

v + (-v) = [0,0,0]

testVI2 :: Int -> Int -> Int -> Bool

Check if simplifyV satisfies:

v * 0 = 0

2. Test for Expr [Integer]

sampleVectorExprIg1 :: Expr [Integer]

a sample vector expression over [Intger]

x + y

sampleVectorExprIg2 :: Expr [Integer]

a sample vector expression over [Integer]

x * y

testVIg1 :: Integer -> Integer -> Integer -> Bool

Check if the expression satisfies:

v + (-v) = [0,0,0]

testVIg2 :: Integer -> Integer -> Integer -> Bool

Check if simplifyV satisfies:

v * 0 = 0

3. Test for Expr [Double]

sampleVectorExprD1 :: Expr [Double]

a sample vector expression over [Double]

x + y

sampleVectorExprD2 :: Expr [Double]

a sample vector expression over [Double]

x * y

testVD1 :: Double -> Double -> Double -> Bool

Check if the expression satisfies:

v + (-v) = [0,0,0]

testVD2 :: Double -> Double -> Double -> Bool

Check if simplifyV satisfies:

v * 0 = 0

4. Test for Expr [Float]

sampleVectorExprF1 :: Expr [Float]

a sample vector expression over [Float]

x + y

sampleVectorExprF2 :: Expr [Float]

a sample vector expression over [Float]

x * y

testVF1 :: Float -> Float -> Float -> Bool

Check if the expression satisfies:

v + (-v) = [0,0,0]

testVF2 :: Float -> Float -> Float -> Bool

Check if simplifyV satisfies:

v * 0 = 0