ExprVector

Description

This module uses the Expr datatype, and UniversalNum class. It contains a type class and instances for vector expressions, and can conduct evaluation, simplicication to those expressions.

Vector is represented as a LIST of numbers, and the types of numbers are Double, Float, Int, Integer respectively.

NOTE: Partial Differentiation , Exponentation and arbitary based Logarithm are not available in the Vector Space.

Class VectorSpace

class Eq a => VectorSpace a where

This class has methods over the Expr datatype that assist with construction and evaluation of vector expressions

Minimal complete definition

evalV, simpV

Methods

evalV :: Map String a -> Expr a -> a

takes a dictionary of variable identifiers and values, and uses it to compute the Expr fully

simpV :: Map String a -> Expr a -> Expr a

takes a possible incomplete dictionary and uses it to reduce Expr as much as possible

simplifyV :: Map String a -> Expr a -> Expr a

simpVlify the expression recursive times untill it can't be reduced anymore

(?+) :: Expr a -> Expr a -> Expr a

Binary Addition & simpVlification

>>> (valV [1,2,3.0] ?+ valV [1,2,3]) :: Expr [Float]
val [2.0,4.0,6.0]

(?*) :: Expr a -> Expr a -> Expr a

Binary production & simplification

The Multiplication of two vectors of same length is a new vector with enties equal to the production of the entries from the 2 vectors in the corresponding position.

(?^) :: Expr a -> a -> Expr a

Exponentiation & simplification

logaV :: a -> Expr a -> Expr a

Logarithm of arbitory base & simplification

sineV :: Expr a -> Expr a

Unary Sin op & simplification

cosiV :: Expr a -> Expr a

Unary Cos & simplification

lnV :: Expr a -> Expr a

Unary Ln & simplification

valV :: a -> Expr a

Wrapper for simple values

varV :: String -> Expr a

String identifier for variables