UniverNumType

Description

The module works with Expr datatype.

The class in the module makes several calculation methods work with all Num type. It expands the domain of functions sin, cos ,ln , log, exponent, 1/ln to both Floating and Integral types. It's a subclass of DiffExpr and VectorSpace classes to support them work with all Num type.

The instances contain Double, Float, Int, Integer 4 basic numerical types.

Class UniverNumType: Univerversal Numerical Type

class (Num a, Eq a) => UniversalNum a where

This class has universal methods for calcuating sin, cos, natural log, log, exponent, and 1/(ln a) over both Floating and Integral numbers.

Minimal complete definition

universalSin, universalCos, universalLn, universalExp, universalLog, universalLogDiffCoe

Methods

universalSin :: a -> a

universalCos :: a -> a

universalLn :: a -> a

universalExp :: a -> a -> a

universalLog :: a -> a -> a

universalLogDiffCoe :: a -> a

the coefficient part of the derivatice of log a e ,i.e. 1/ln a